updated to version 2.2
changelog:

-New Module: GLTCH3HR

GLTCH3HR is a threshold-triggered looper which, upon receiving a signal that exceeds the threshold, records 0.5s of a given input, and replays a section of it defined by the start and end parameters until the signal goes below the threshold. The start and end can be freely manipulated during the looping period, and can even reverse the playback in realtime. More information in the readme.

FM3HR:
-Added anti-aliasing to the base waveforms.
-Removed sinesaw waveform.
-Removed reset input, as it wasn't implemented anyway.

Repeat3hr:
-Fixed bug where time parameter could go below 0, causing the algorithms to crash.